// src/modules/invoice/usecase/find/invoice.dto.ts
export interface FindInvoiceUseCaseInputDto {
    id: string;
}

export interface FindInvoiceUseCaseOutputDto {
    id: string;
    name: string;
    document: string;
    address: {
        street: string;
        number: string;
        complement: string;
        city: string;
        state: string;
        zipCode: string;
    };
    items: {
        id: string;
        name: string;
        price: string;
    }[];
    total: number;
    createdAt: Date;
}